+++ travis-ui.el
@@ -25,6 +25,11 @@
 (require 'tabulated-list)
 
 
+;;
+;; to display faces in IELM :
+;; ELISP (list-faces-display "travis")
+;;
+
 ;; Faces
 
 (defface travis--title
@@ -34,8 +39,18 @@
   :group 'travis)
 
 (defface travis--gray-face
-  '((((class color)) :foreground "#"))
+  '((((class color)) :foreground "#b1b6b6"))
   "Gray color.."
+  :group 'travis)
+
+(defface travis--cyan-face
+  '((((class color)) :foreground "#00ffff"))
+  "Cyan color.."
+  :group 'travis)
+
+(defface travis--yellow-face
+  '((((class color)) :foreground "#e5e500"))
+  "Yellow color."
   :group 'travis)
 
 (defface travis--orange-face
@@ -53,21 +68,21 @@
   "Green color."
   :group 'travis)
 
-(defface travis--yellow-face
-  '((((class color)) :foreground "#e5e500"))
-  "Yellow color."
-  :group 'travis)
 
 
 (defun colorize-build-state (state)
   "Colorize face using STATE."
   (cond
+   ((string= state "canceled")
+    (propertize state 'face 'travis--gray-face))
    ((string= state "failed")
     (propertize state 'face 'travis--red-face))
    ((string= state "errored")
     (propertize state 'face 'travis--orange-face))
    ((string= state "passed")
     (propertize state 'face 'travis--green-face))
+   ((string= state "created")
+    (propertize state 'face 'travis--cyan-face))
    (t (propertize state 'face 'travis--gray-face))))