+++ ChangeLog.md
@@ -1,9 +1,11 @@
 # emacs-travis ChangeLog
 
-# Version 0.2.0 (IN PROGRESS)
+# Version 0.2.0 (11/18/2014)
 
+- Test TravisUI faces
 - Implements Travis mode to display projects build status
 - Enable code coverage
+- Configure travis, drone.io for continuous integration
 
 # Version 0.1.0 (11/14/2014)